Pairing .ipynb file with a .py file for Version Control
In the terminal we pair the .ipynb file ocntaining all of the code with a .py file which will be kept in sync. The latter will be used to commit changes to the Git repository for the project. This way there will be no need every time changes to the .ipynb file are made for it to be downloaded and the old version manually replaced in the directory.